requirements.txt
pygame-ce
pytmx-ng
beautifulsoup4
lxml
astar
blinker
pygame_gui
ruff
